+++ src/main/java/racingcar/controller/GameController.java
@@ -1,5 +1,7 @@
-package racingcar.domain;
+package racingcar.controller;
 
+import racingcar.domain.Car;
+import racingcar.service.GameService;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
