+++ src/main/java/racingcar/service/GameService.java
@@ -1,6 +1,7 @@
-package racingcar.domain;
+package racingcar.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
+import racingcar.domain.Car;
 
 import java.util.List;
 import java.util.stream.Collectors;
